+++ lib/ayumi_web/router.ex
@@ -70,9 +70,11 @@ defmodule AyumiWeb.Router do
       live "/support_records", SupportRecordLive.Index, :index
       live "/service_users/:service_user_id/attendance", AttendanceLive.Index, :index
       live "/service_users/:service_user_id/attendance/sheet", AttendanceLive.Sheet, :index
+      live "/exports", ExportLive.Index, :index
     end
 
     post "/users/update-password", UserSessionController, :update_password
+    get "/exports/download", ExportController, :download
   end
 
   scope "/", AyumiWeb do